CodeSphere/requirements.txt
flask
python-dotenv
anthropic
pytesseract
opencv-python
numpy
Pillow
